Type conversion hataları düzeltildi.

+++ src/main/java/com/tobeto/bootcampProject/dataacces/ApplicantRepository.java
@@ -3,5 +3,5 @@ package com.tobeto.bootcampProject.dataacces;
 import com.tobeto.bootcampProject.entities.Applicant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicantRepository extends JpaRepository<Applicant, Long> {
+public interface ApplicantRepository extends JpaRepository<Applicant, Integer> {
 }
