+++ src/main/java/com/tobeto/bootcampProject/dataacces/InstructorRepository.java
@@ -3,5 +3,5 @@ package com.tobeto.bootcampProject.dataacces;
 import com.tobeto.bootcampProject.entities.Instructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InstructorRepository extends JpaRepository<Instructor, Long> {
+public interface InstructorRepository extends JpaRepository<Instructor, Integer> {
 }
